Refactor: move def parsing into separate class and add tests
